Change Rectangle.area to long

+++ basic_programming/lesson_49/code/date/src/homework_48/Rectangle.java
@@ -6,7 +6,7 @@ public class Rectangle {
   private final int width; // ширина // неизменяемое значение, хотите новое - создайте новый объект
 //  private final int height; // обычно в программировании ширина и высота, но у нас геометрия,
   // где длина и ширина
-  private final int area; // можно считать каждый раз при вызове метода, а можно - заранее;
+  private final long area; // можно считать каждый раз при вызове метода, а можно - заранее;
   // нам точно выгоднее заранее, потому что все наши размеры - final и не будут меняться.
 
   public Rectangle(int a, int b) {
@@ -14,7 +14,10 @@ public class Rectangle {
     // это геометрия, а не дизайн, тогда прямоугольники можно вращать без последствий
     length = Math.max(a, b);
     width = Math.min(a, b);
-    area = a * b;
+//    if ((long) a * (long) b != a * b) { // если площадь - int и мы хотим проверить пределы
+//      throw new RuntimeException("The area is too big for the sides: " + a + ", " + b);
+//    }
+    area = (long) a * (long) b;
   }
 
   public int getLength() {
@@ -25,7 +28,7 @@ public class Rectangle {
     return width;
   }
 
-  public int getArea() {
+  public long getArea() {
 //    return length * width; // можно считать именно здесь, каждый раз заново
     return area;
   }
